refactor: 테스트 계정으로 로그인 하는 과정 코드를 수정하였습니다
