+++ core/domain/src/main/java/com/teampatch/core/domain/repository/AuthenticationRepository.kt
@@ -5,6 +5,4 @@ import com.teampatch.core.domain.model.LoginResult
 interface AuthenticationRepository {
 
     suspend fun loginKakao(): LoginResult
-
-    suspend fun loginTestAccount(): LoginResult
 }
