--- core/domain/src/main/java/com/teampatch/core/domain/usecase/authentication/LoginTestAccountUseCase.kt
@@ -1,13 +1,0 @@
-package com.teampatch.core.domain.usecase.authentication
-
-import com.teampatch.core.domain.repository.AuthenticationRepository
-import javax.inject.Inject
-
-class LoginTestAccountUseCase @Inject constructor(
-    private val authenticationRepository: AuthenticationRepository,
-) {
-
-    suspend operator fun invoke() {
-        authenticationRepository.loginTestAccount()
-    }
-}
